% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mids.R
\name{compare_mids}
\alias{compare_mids}
\title{Compare Two \code{mids} Objects and Show Summary}
\usage{
compare_mids(
  x,
  y,
  show_msg = TRUE,
  show_rtn = !show_msg,
  ignore_rng = NULL,
  ignore_call = TRUE,
  ignore_version = TRUE,
  ignore_date = TRUE
)
}
\arguments{
\item{x, y}{\code{mids} objects to compare, as created by
\code{\link[mice:mice]{mice::mice()}},
\code{\link[futuremice:future_mice]{future_mice()}},
\code{\link[mice:mice.mids]{mice::mice.mids()}},
\code{\link[futuremice:future_mids]{future_mids()}}, etc.}

\item{show_msg}{Should the message displaying element-wise comparisons be
shown?}

\item{show_rtn}{Should the return value be visible? The default is the
opposite of \code{show_msg}, so that either the message or return value is
shown, but not both.}

\item{ignore_rng}{Should elements or components of elements affected by the
RNG state be ignored? See Details for more information. The default is to
ignore when different package namespaces are detected in the \code{mids$call},
otherwise to check.}

\item{ignore_call, ignore_version, ignore_date}{Should the corresponding
\code{mids} element (\code{mids$call}, \code{mids$version}, \code{mids$date}, respectively) be
ignored in equality comparison? These elements do not usually impact
results if all other elements are identical and can thus usually be safely
ignored.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{compare_mids()} is a combination of
\code{\link[base:all.equal]{all.equal()}} and
\code{\link[base:identical]{identical()}} for \code{mids} objects. Like
\code{all.equal()}, it displays a summary of element-wise differences (it also
explicitly lists elements that are equal). However, rather than outputting
these differences as a \code{character} vector if found, it always returns either
\code{TRUE} or \code{FALSE}. If \code{quiet = FALSE}, the return value is invisible.
}
\details{
Proper random number generation requires different methods when performing
operations in parallel; thus, is it generally not possible to exactly
replicate \code{mids} objects produced sequentially when imputing chains in
parallel. In particular, the imputed values will differ, as well as any
functions of those values (\code{chainMean}, \code{chainVar}, \code{loggedEvents}). This
function takes the stance that imputations performed by functions in the
same package should be consistent, but functions from different packages
may handle random number generation differently. Thus, if \code{compare_mids()}
can detect that two functions came from the same package, it will compare
the full RNG-related elements by default. Otherwise, it will check that
attributes independent of random number generation are the same, but it will
allow differences between attributes and data dependent on the RNG state. If
RNG-related attributes are ignored, the RNG state itself is also ignored, so
differences in \code{seed} and \code{lastSeedValue} are also allowed in this case.
}
