% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mice.R
\name{fm_caller_args}
\alias{fm_caller_args}
\title{Construct List of Arguments from Matched Caller Function}
\usage{
fm_caller_args(
  n = 1,
  ...,
  defaults = FALSE,
  dots_env = NULL,
  dots_expand = TRUE
)
}
\arguments{
\item{n}{The number of callers to go back.}

\item{...}{These dots must be empty.}

\item{defaults}{Whether to match missing arguments to their
defaults.}

\item{dots_env}{An execution environment where to find dots. If
supplied and dots exist in this environment, and if \code{call}
includes \code{...}, the forwarded dots are matched to numbered dots
(e.g. \code{..1}, \code{..2}, etc). By default this is set to the empty
environment which means that \code{...} expands to nothing.}

\item{dots_expand}{If \code{FALSE}, arguments passed through \code{...} will
not be spliced into \code{call}. Instead, they are gathered in a
pairlist and assigned to an argument named \code{...}. Gathering dots
arguments is useful if you need to separate them from the other
named arguments.

Note that the resulting call is not meant to be evaluated since R
does not support passing dots through a named argument, even if
named \code{"..."}.}
}
\value{
A named list of arguments.
}
\description{
Construct List of Arguments from Matched Caller Function
}
\keyword{internal}
