% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_utils.R
\name{fm_exit_msg}
\alias{fm_exit_msg}
\title{Throw Messages/Warnings at End of \code{future_mice()} Execution}
\usage{
fm_exit_msg(i, rhat, minit, rhat_msg)
}
\arguments{
\item{i}{Integer(ish) representing iteration count}

\item{rhat_msg}{Contents of \code{message} displaying R-hat values for last
\code{rhat_it} iterations}

\item{rhat_lt}{Logical vector of R-hat comparisons. \code{length(rhat_lt)} must be
less than or equal to \code{rhat_it}.}

\item{rhat_it}{Integer(ish) number of iterations used in R-hat comparison}
}
\value{
\code{NULL}, invisibly
}
\description{
Throw Messages/Warnings at End of \code{future_mice()} Execution
}
\keyword{internal}
