% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_utils.R
\name{fm_furrr_opts}
\alias{fm_furrr_opts}
\title{Create \code{furrr_options()} List from Parallelization Parameters}
\usage{
fm_furrr_opts(parallel_params)
}
\arguments{
\item{parallel_params}{List of parameters for parallelization as calculated
by \code{fm_parallel_params()}}
}
\value{
A list of options for \code{furrr} functions, as created by
\code{furrr::furrr_options()}
}
\description{
Create \code{furrr_options()} List from Parallelization Parameters
}
\keyword{internal}
