% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_utils.R
\name{fm_gcd}
\alias{fm_gcd}
\title{Calculate Greatest Common Divisor of Positive Integers}
\usage{
fm_gcd(...)
}
\arguments{
\item{...}{Numeric vectors containing integer set for GCD calculation}
}
\value{
A positive scalar \code{integer} containing the GCD of the inputs
}
\description{
Calculate Greatest Common Divisor of Positive Integers
}
\keyword{internal}
