% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibindlist.R
\name{fm_ibind_logged_events}
\alias{fm_ibind_logged_events}
\title{Helper Function for Setting \code{mids$loggedEvents}}
\usage{
fm_ibind_logged_events(mids, mids_list)
}
\arguments{
\item{mids}{A \code{mids} object to update}

\item{mids_list}{A list of \code{mids} objects with \code{loggedEvents} to extract and
combine}
}
\value{
\code{mids} with \code{loggedEvents} from \code{mids_list}
}
\description{
Helper Function for Setting \code{mids$loggedEvents}
}
\keyword{internal}
