% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_assert.R
\name{fm_is_scalar_int}
\alias{fm_is_scalar_int}
\title{Is Object a Scalar Integer-ish Value}
\usage{
fm_is_scalar_int(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Always returns \code{TRUE} or \code{FALSE} - never errors
}
\keyword{internal}
