% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isplit.R
\name{fm_isplit_logged_events}
\alias{fm_isplit_logged_events}
\title{Subset and Insert \code{mids$loggedEvents} Element}
\usage{
fm_isplit_logged_events(mids, logged_events, im)
}
\arguments{
\item{mids}{A \code{mids} object to insert new \code{loggedEvents} into}

\item{logged_events}{A \code{data.frame} containing the event logs from a
\code{mice::mice()} run}

\item{im}{The imputations for which the associated logs will be inserted}
}
\value{
\code{mids} with a new \code{loggedEvents} element
}
\description{
Subset and Insert \code{mids$loggedEvents} Element
}
\keyword{internal}
