% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mice.R
\name{fm_mice}
\alias{fm_mice}
\title{\code{{furrr}}-Friendly \code{mice::mice()} w/ Progress Updates}
\usage{
fm_mice(.m, mice_args, progressor)
}
\arguments{
\item{.m}{Sink that allows iteration w/ \code{{purrr}}-style \code{map()} functions.
Unused.}

\item{progressor}{A \code{progressor()} from \code{{progressr}}}

\item{...}{Arguments passed on to \code{mice::mice}. \code{seed} is ignored.}
}
\value{
A \code{mids} object (\emph{m}ultiply \emph{i}mputed \emph{d}ata \emph{s}et)
}
\description{
\code{{furrr}}-Friendly \code{mice::mice()} w/ Progress Updates
}
\keyword{internal}
