% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mice.R
\name{fm_mice_args}
\alias{fm_mice_args}
\title{Combine Arguments in \code{future_mice()}}
\usage{
fm_mice_args(m, .args = fm_caller_args(n = 2L))
}
\arguments{
\item{m}{The number of chains per \code{mice::mice()} call}

\item{.args}{A named list of arguments from the \code{future_mice()} call}
}
\value{
A list containing arguments to pass to \code{mice::mice()}
}
\description{
Helper function to combine and parse named arguments + dots in
\code{future_mice()}
}
\keyword{internal}
