% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fm_mice_seed}
\alias{fm_mice_seed}
\title{Create \code{{mice}}-Friendly \code{seed}}
\usage{
fm_mice_seed(seed)
}
\arguments{
\item{seed}{A scalar \code{integer}, \code{NA}, or \code{NULL}}
}
\value{
Converts \code{NULL} to \code{NA}, otherwise returns input or errors
}
\description{
Create \code{{mice}}-Friendly \code{seed}
}
\keyword{internal}
