% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mids.R
\name{fm_mids}
\alias{fm_mids}
\title{\code{{furrr}}-Friendly \code{mice::mice.mids()} w/ Progress Updates}
\usage{
fm_mids(mids, mice_args, progressor, newdata = NULL)
}
\arguments{
\item{mids}{An object of class \code{mids}, typically produced by a previous call
to \code{fm_mice()}}

\item{progressor}{A \code{progressor()} from \code{{progressr}}}

\item{newdata}{An optional \code{data.frame} for which multiple imputations are
generated according to the model in \code{obj}}

\item{maxit}{The number of additional Gibbs sampling iterations}
}
\value{
A \code{mids} object (\emph{m}ultiply \emph{i}mputed \emph{d}ata \emph{s}et) with additional
iterations
}
\description{
\code{{furrr}}-Friendly \code{mice::mice.mids()} w/ Progress Updates
}
\keyword{internal}
