% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_utils.R
\name{fm_parallel_params}
\alias{fm_parallel_params}
\title{Calculate Parameters for Parallelization of Chains}
\usage{
fm_parallel_params(m, chunk_size, maxit, seed)
}
\arguments{
\item{m}{The total number of chains (imputations)}

\item{chunk_size}{The average number of chains per future}
}
\value{
A \code{list} containing parameters \code{n_chains} (chains per call),
\code{n_calls} (number of calls to \code{mice()}),
\code{chunk_size} (number of calls per future),
\code{maxit} (maximum number of iterations), and \code{seed} (RNG seed)
}
\description{
Calculates the number of chains per call to \code{mice::mice()} or
\code{mice::mice.mids()}, the average number of calls chunked into a future, and
the number of futures needed to satisfy the given arguments.
}
\keyword{internal}
