% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat_max.R
\name{fm_prep_diagnostic_params}
\alias{fm_prep_diagnostic_params}
\title{Extract Chain Means and Variances from a \code{mids} object}
\usage{
fm_prep_diagnostic_params(mids)
}
\arguments{
\item{mids}{A \code{mids} object as created by \code{mice::mice()}}
}
\value{
A list of 2D matrices (one for each imputed variable) with rows
corresponding to iterations and columns corresponding to chains
}
\description{
Extract Chain Means and Variances from a \code{mids} object
}
\keyword{internal}
