% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_utils.R
\name{fm_progressor}
\alias{fm_progressor}
\title{Initialize \code{progressor} for \code{future_mice()} and \code{future_mids()}}
\usage{
fm_progressor(parallel_params, progressor = NULL)
}
\arguments{
\item{parallel_params}{List of parameters for parallelization as calculated
by \code{fm_parallel_params()}}

\item{progressor}{An optional existing \code{progressor} object, or \code{NULL}}
}
\value{
A \code{progressor} object; if \code{progressor != NULL}, this is the unchanged
input
}
\description{
Initialize \code{progressor} for \code{future_mice()} and \code{future_mids()}
}
\keyword{internal}
