% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat_max.R
\name{fm_rhat_converged}
\alias{fm_rhat_converged}
\title{Check Convergence of \code{mids} Object}
\usage{
fm_rhat_converged(mids, n = 1L, max = 1.05)
}
\arguments{
\item{mids}{A \code{mids} object with imputations to check}

\item{n}{The number of iterations to use when checking. Must be > 0.}

\item{max}{The upper bound for R-hat convergence}
}
\description{
Check Convergence of \code{mids} Object
}
