% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat_max.R
\name{fm_rhat_max_}
\alias{fm_rhat_max_}
\title{Quick Max R-hat Calculation}
\usage{
fm_rhat_max_(params, iters = NULL)
}
\arguments{
\item{iters}{An \code{integer} vector of iterations to use in the R-hat
R-hat calculation}
}
\value{
A scalar \code{double} containing the maximum R-hat statistic
}
\description{
Calculates the largest R-hat statistic across all variables and chain
statistics for the most recent iteration
}
\keyword{internal}
