% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fm_rng_seed}
\alias{fm_rng_seed}
\title{Create \code{{rngtools}}-Friendly \code{seed}}
\usage{
fm_rng_seed(seed)
}
\arguments{
\item{seed}{A scalar \code{integer}, \code{NA}, or \code{NULL}}
}
\value{
Converts \code{NA} to \code{NULL}, otherwise returns input or errors
}
\description{
Create \code{{rngtools}}-Friendly \code{seed}
}
\keyword{internal}
