% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibindlist.R
\name{fm_set_call}
\alias{fm_set_call}
\title{Helper Function for Setting \code{mids$call}}
\usage{
fm_set_call(mids, call)
}
\arguments{
\item{mids}{A \code{mids} object to update}

\item{call}{A \code{call} object to use as \code{mids$call} after standardization. Can
be \code{NULL}, in which case \code{mids} is returned unaltered.}
}
\value{
The \code{mids} object with an updated \code{call} (if \code{call} is not \code{NULL}).
Note that if \code{call} has arguments \code{m}, \code{maxit}, or \code{seed}, their values
are updated to reflect the corresponding attribute in \code{mids}.
}
\description{
Helper Function for Setting \code{mids$call}
}
\keyword{internal}
