% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fm_set_colnames}
\alias{fm_set_colnames}
\title{Helper Function for Setting Column Names}
\usage{
fm_set_colnames(x, names)
}
\arguments{
\item{x}{An object to set column names for. Must have at least 2 dimensions.}

\item{names}{A character vector of column names}
}
\value{
\code{x}, with (re-)named columns
}
\description{
Helper Function for Setting Column Names
}
\keyword{internal}
