% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_utils.R
\name{fm_set_names}
\alias{fm_set_names}
\alias{fm_set_colnames}
\alias{fm_set_rownames}
\title{Helper Function for Setting Row and Column Names}
\usage{
fm_set_colnames(x, names)

fm_set_rownames(x, names)
}
\arguments{
\item{x}{An object to set names for. Must have at least 2 dimensions to use
\code{fm_set_colnames()}.}

\item{names}{A \code{character} vector of row or column names}
}
\value{
\code{x}, with (re-)named rows or columns
}
\description{
Helper Function for Setting Row and Column Names
}
\keyword{internal}
