% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mids.R
\name{future_mids}
\alias{future_mids}
\title{Parallelize \code{\link[mice:mice.mids]{`mice::mice.mids()`}}  Using \code{{future}}}
\usage{
future_mids(
  obj,
  newdata = NULL,
  maxit = 1L,
  quiet = FALSE,
  chunk_size = 1L,
  rhat_thresh = 1.05,
  rhat_it = 3L,
  progressor = NULL,
  update_call = TRUE,
  ...
)
}
\arguments{
\item{obj}{A \code{mids} object, as created by
\code{\link[mice:mice]{mice::mice()}},
\code{\link[futuremice:future_mice]{future_mice()}},
or \code{future_mids()} (this function)}

\item{newdata}{An optional \code{data.frame} for which multiple imputations
are generated according to the model in \code{obj}.}

\item{maxit}{The number of additional Gibbs sampling iterations.}

\item{chunk_size}{The average number of chains per future. Differs from the
usual \code{{future}} parameter in that multiple chains ("chunks") will be
evaluated in a single call to \code{mice::mice()} if there is an integer \code{i}
such that \verb{1 < i <= chunk_size} and \code{m \%\% i == 0}.}

\item{rhat_thresh}{The R-hat threshold used to assess convergence.
Convergence is defined as \code{all(tail(rhat, rhat_it) < rhat_thresh)}.}

\item{rhat_it}{The number of iterations used to assess convergence.
Convergence is defined as \code{all(tail(rhat, rhat_it) < rhat_thresh)}.}

\item{progressor}{An optional \code{\link[progressr]{progressor}}
function to signal progress updates. If supplied, you are responsible for
ensuring that the number of steps in the \code{progressor} is consistent with
the number of iterations performed in \code{future_mice()}.}

\item{update_call}{Should \code{mids$call} be set to new \code{future_mids()} call or
left unchanged?}

\item{...}{Named arguments that are passed down to the univariate imputation
functions.}
}
\value{
Returns an S3 object of class \code{\link[mice:mids-class]{mids}}
       (multiply imputed data set)
}
\description{
\code{future_mids()} is analogous to \code{mice::mice.mids()}, but parallelizes chains
using the \code{{furrr}} package and stops early if convergence is detected using
the R-hat statistic
(see the \code{\link[futuremice:future_mice]{future_mice()}} documentation
for details).
}
