% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futuremice-package.R
\docType{package}
\name{futuremice-package}
\alias{futuremice}
\alias{futuremice-package}
\title{futuremice: A Future for MICE}
\description{
futuremice provides parallelized variants of `mice::mice()` and `mice::mice.mids()` which support progress updates and early stopping on convergence.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jesse-smith/futuremice}
  \item \url{https://jesse-smith.github.io/futuremice/}
  \item Report bugs at \url{https://github.com/jesse-smith/futuremice/issues}
}

}
\author{
\strong{Maintainer}: Jesse Smith \email{jesse.smith@stjude.org}

}
\keyword{internal}
