% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibindlist.R
\name{ibindlist}
\alias{ibindlist}
\title{Combine a List of \code{\link[mice]{mids}} Objects}
\usage{
ibindlist(mids_list, call = NULL, seed = NULL, last_seed_value = NULL)
}
\arguments{
\item{mids_list}{List of \code{mids} objects to combine}

\item{call}{An optional call to use for the \code{call} attribute of the resulting
\code{mids} object. The default uses the call in \code{mids_list[[1]]$call}.}

\item{seed}{An optional integer to set as the \code{seed} attribute of the
resulting \code{mids} object.}

\item{last_seed_value}{An optional \code{\link[base:Random]{.Random.seed}}
value to set as the \code{lastSeedValue} attribute of the resulting \code{mids}
object}
}
\value{
A combined \code{mids} object
}
\description{
Combines a list of \code{\link[mice]{mids}} objects into a single \code{mids}
object. The resulting number of imputed data sets is equal to the sum of the
number of imputed data sets in each list element.
}
\details{
The \code{call}, \code{seed}, and \code{last_seed_value} arguments are primarily used by
\code{\link[futuremice:future_mice]{future_mice()}} and
\code{\link[futuremice:future_mids]{future_mids()}}; they allow modification
of the \code{mids} object to match the equivalent
\code{\link[mice:mice]{mice::mice()}} output exactly.
}
\examples{
# Create individual `mids` objects
mids1 <- mice::mice(mice::nhanes, m = 1L, maxit = 1L, seed = 1L)
mids2 <- mice::mice(mice::nhanes, m = 2L, maxit = 1L, seed = 2L)
mids3 <- mice::mice(mice::nhanes, m = 3L, maxit = 1L, seed = 3L)

# Combine imputations
mids6 <- ibindlist(list(mids1, mids2, mids3))
mids6

}
