% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat_max.R
\name{prep_diagnostic_params}
\alias{prep_diagnostic_params}
\title{Extract Chain Means and Variances from a \code{mids} object}
\usage{
prep_diagnostic_params(data)
}
\arguments{
\item{data}{An object of class \code{mids} as created by \code{mice::mice()}}
}
\value{
A list of 2D matrices (one for each imputed variable) with rows
corresponding to iterations and columns corresponding to chains
}
\description{
Extract Chain Means and Variances from a \code{mids} object
}
\keyword{internal}
