% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat_max.R
\name{rhat_max}
\alias{rhat_max}
\title{Quick Max R-hat Calculation}
\usage{
rhat_max(data)
}
\arguments{
\item{data}{An object of class \code{mids} as created by \code{mice::mice()}}
}
\value{
A legnth 1 double containing the maximum R-hat statistic
}
\description{
Calculates the largest R-hat statistic across all variables and chain
statistics for the most recent iteration
}
\keyword{internal}
