% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat_max.R
\name{rstan_rhat}
\alias{rstan_rhat}
\alias{rstan_rhat_}
\alias{rstan_split_chains}
\alias{rstan_z_scale}
\alias{rstan_is_constant}
\title{Re-Implement \code{rstan::Rhat()}}
\usage{
rstan_rhat(params)

rstan_rhat_(params)

rstan_split_chains(params)

rstan_z_scale(x)

rstan_is_constant(x, tol = .Machine$double.eps)
}
\arguments{
\item{params}{A 2D matrix or array with one row per iteration and one column
per chain. The cells are realized draws for a particular parameter or
function of parameters.}

\item{x}{Vector, matrix, or array to scale}
}
\value{
\code{rstan_rhat()} returns a scalar \code{double} that is the max of the bulk
and tail (folded) R-hat statistic for \code{params}. \code{rstan_rhat_()} returns a
scalar \code{double} R-hat. \code{rstan_split_chains()} returns an array with each
column split into two columns (top and bottom half). \code{rstan_z_scale()}
returns the scaled vector/matrix/array.
}
\description{
Internal implementation of \code{rstan::Rhat()} to avoid unnecessary dependency.
}
\keyword{internal}
