% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_mice.R
\name{fm_mice}
\alias{fm_mice}
\title{\code{{furrr}}-Friendly \code{mice::mice()} w/ Progress Updates}
\usage{
fm_mice(.m, p, mice_args)
}
\arguments{
\item{.m}{Sink for using \code{{purrr}}-style \code{map()} functions for iteration.
Unused.}

\item{p}{\verb{[progressor]} A progress bar object from \code{{progressr}}}

\item{...}{Arguments passed on to \code{mice::mice}. \code{seed} is ignored.}
}
\value{
A \code{mids} object (\emph{m}ultiply \emph{i}mputed \emph{d}ata \emph{s}et)
}
\description{
\code{{furrr}}-Friendly \code{mice::mice()} w/ Progress Updates
}
\keyword{internal}
