% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat_max.R
\name{rhat_max}
\alias{rhat_max}
\title{Quick Max R-hat Calculation}
\usage{
rhat_max(mids, n = 1L)
}
\arguments{
\item{mids}{A \code{mids} object as created by \code{mice::mice()}}

\item{n}{The number of recent iterations for which R-hat should be
calculated. If this is larger than the total number of iterations, it is
truncated.}
}
\value{
A scalar \code{double} containing the maximum R-hat statistic
}
\description{
Calculates the largest R-hat statistic across all variables and chain
statistics for the \code{it} most recent iterations
}
\examples{
# Create `mids` object
mids <- mice::mice(mice::nhanes)

# Get max R-hat for most recent 2 iterations
rhat <- rhat_max(mids, it = 2L)
rhat

}
